Add test assertions
